work_with_database/requirements.txt
django~=4.0.3
psycopg2-binary

django-autoslug~=1.9.8